got gps working on Android
Note, to get this to work I manually edited AndroidManifest.xml which might get blown out by a future Meteor build and the xml file appears to be under the .gitignore list. Not sure how this normally gets set but on another project I didn't have to set it manually – maybe removing and re-adding the geolocation cordova plugin would add it automatically. Not sure why it didn't the first time.