migrated form error messages into db
messages now have a tplKey which is the str short form for retrieval and the full message is tokenized where each placeholder (aka token) is in the form of "{#}" where # is the index of the array of replacement values provided. Created a filter for angular templates to be used like {{"Hello {0}, how are you? {0} is a nice name, my name is {1}" | tokenValues:['Ernie', 'Bert']}}