bug fix: disappearing back button
The problem was that both the bottom of the AppCtrl and the Accounts.onLogin handler were calling startup (which calls $state.go('app.home')). Ionic's history breaks when you $state.go to a state that you're already on. Basically there was a sort of redundant race condition. I have made the onLogin handler specific to the user logging in through the login and create account views. When browser reloads (manually or via hot push) previously logged in users are automatically logged in. This used to trigger the Accounts.onLogin handler. This no longer happens for the reload use case.